refactor(host-agent): extract crates/shed-broker — the embeddable broker core
Move the broker core (bus, supervisor, watcher, discovery, config, ssh/aws/
docker backends, egress, minter, bootstrap, controltoken, approval, audit,
touchid, LiveStatus snapshot, socket paths+probes) out of the shed-host-agent
bin into a new lib crate, leaving the daemon a thin shell (CLI, signals,
run_daemon wiring, status UDS server, Surface-A desktop server, the
desktop-forwarding feature). Groundwork for 3a.2: the desktop app embeds this
lib in-process so no separate host-agent install is needed; the headless
standalone daemon persists unchanged.

Four seam inversions beyond the move (plan §3.1):
- ControlTokenMinter + MintedControlToken relocate from desktop.rs into the
  lib's controltoken module (un-gated — embedders need token minting).
- audit's cfg'd Option<Arc<DesktopServer>> fan-out becomes an always-compiled
  injectable AuditFanout seam; one JsonlAuditSink constructor; the
  AuditEntryView conversion moves bin-side.
- sockets splits: path resolution + liveness probes (pub) in the lib; the
  bind ceremony + Log coupling stay bin-side (socket_bind.rs).
- status splits: snapshot builder in the lib; UDS server + status CLI client
  bin-side (status_server.rs). select_builtin_gate hoists policy→gate routing
  (None = shed-desktop, caller composes), preserving exact routing in both
  feature configs.

Wire-behavior invariant held: 108/108 Go-vs-Rust differential cells green on
the rebuilt binary; headless --no-default-features leg green; clippy -D
warnings clean workspace-wide; make check (Go) green; goreleaser --package
shed-host-agent wiring unchanged. ci.yml hostagent filter learns
crates/shed-broker/**; crates/CLAUDE.md carve-out notes updated.

tests/golden.rs rewired from a #[path] include of config.rs to the lib's
public try_parse (same coverage; the #[cfg(test)] parse helper is unavailable
to an external integration test).

Review (cursor, adversarial, correctness-scoped): no real bugs; one accepted
API-surface note — the now-pub HostAgentConfig.discovery field is mutable
post-load by embedders without re-validate; acceptable, the only planned
embedder constructs configs through the same load paths (plan §3.1
minimal-pub). Simplify pass: MINIMAL (three pre-existing stale comments in
moved-verbatim bodies noted for a later non-differential cleanup).
