+++ crates/shed-broker/src/supervisor.rs
